move gmab_options to a module and add tests
