+++ gmab/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod arm;
 mod genetic;
 pub mod gmab;
+pub mod gmab_options;
 mod sorted_multi_map;
